Add initial blocks support
View blocks are reusable sets of VNode lifecycle callbacks, think something
like React components with their `render` defined in template files.

To use a block, pass a corresponding block class as a key to `blocks` option
of `render` function. For example:

    var blocks = {
        // key is the block name, value is the block class
        form: FormBlock
    };

    var myRender = function(state) {
        return render(state, h, { blocks: blocks });
    };

Now the `form` block can be used by templates, just add a `BLOCK_NAME` attribute
to `TMPL_INLINE` tag:

    <div class="container">
        <TMPL_INLINE BLOCK_NAME="form" form_content
            action="/login">
    </div>

As an example, `action` attribute (string "/login") is passed to the block -
it will be available as the first argument of `FormBlock` constructor:

    function FormBlock(props) {
        this.props = props;
        console.log(props.action); // logs "/login";
    }

Currently the following lifecycle callbacks are executed (each of these callbacks
is optional):

- `blockWillMount`
- `blockDidMount`
- `blockWillUpdate`
- `blockDidUpdate`
- `blockWillUnmount`
- `shouldBlockUpdate`

These methods and their signatures were copied from React's API and should more
or less correspond to it.